\name{EMVC-package}
\alias{EMVC-package}
\docType{package}
\title{
Implementation of the Entropy Minimization over Variable Clusters (EMVC) algorithm
}
\description{
Contains logic for the data-driven optimization of 
annotations via minimization of the entropy of variable group
members over discrete partitions of variables generated via 
either k-means clustering or horizontal cuts of dendrograms output via
agglomerative hierarchical clustering.
}
\details{
\tabular{ll}{
Package: \tab EMVC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-13\cr
License: \tab GPL-3 \cr
}
Variable group optimization is performed using the function \code{EMVC}.
}
\author{
H. Robert Frost and Jason H. Moore
}
\references{
Frost, HR, Moore, JH. Optimization of gene sets via entropy minimization over variable clusters (EMVC). Submitted.
}
\note{
This work was supported by the National Institutes of Health R01 LM010098.
}
\keyword{ package }

